refactor(inventory): extract grant.rs tests to sibling tests.rs (#244)
grant.rs was 709 lines (just over the 700-line cap). Production code
(handle_grant_item + normalize_item_ids + item_allows_container) is 454
lines and stays one cohesive transaction; the 256 lines of test code
were what pushed the file over the cap.

  grant/mod.rs   — production helpers + handle_grant_item
  grant/tests.rs — normalize_item_ids unit tests + live-DB
                   grant flow tests (require_db_or_skip!)

Public surface unchanged: callers still use
`grant::handle_grant_item`, `grant::normalize_item_ids`, etc.

All 8 grant tests pass (live-DB ones skip cleanly without DATABASE_URL);
clippy clean; cargo fmt applied.
